backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
httpx>=0.28
mistralai>=1.5
anthropic>=0.42
python-dotenv>=1.0
pydantic>=2.10
